Podcast: extract endpoint relay_response into Relay_Response trait
Fixes Phan undeclared-type errors on Podcast_Distribution_Endpoint
(missing WP_REST_Response import) by consolidating the duplicated
relay_response helper across Posts_To_Podcast_Endpoint,
Podcast_Stats_Endpoint, and Podcast_Distribution_Endpoint into a
shared trait.
